some classes and packages renaming
